Extract NavBar
